share reclamm chart btwn pool and creation pages
